SampleCode/requirements.txt
azure-keyvault==1.1.0
